Fix rx_buffer data loss, UB in LongVariable decode, add NACK handling and entity validation
- nasa_client: erase only consumed bytes from rx_buffer instead of clearing
  all, preventing loss of trailing data from the next packet when multiple
  packets queue in the UART FIFO
- nasa_client: handle NACK responses by removing packet from retry queue
  instead of resending until timeout
- nasa_message: use unsigned intermediates in LongVariable decode to avoid
  signed left-shift UB in C++17
- nasa_packet: fix stale comment about frame size (size+3 → size+2)
- const: fix hp_current accuracy from 2 to 1 (divisor 10 gives 0.1A resolution)
- __init__: validate entity-to-device-type compatibility (prevent indoor
  codes on outdoor devices and vice versa)
- dashboard: document required HACS custom card dependencies

+++ components/nasactl/nasa_message.h
@@ -46,10 +46,11 @@ struct MessageSet {
       case MessageSetType::LongVariable:
         if (offset + 4 > data.size())
           return false;
-        value = (static_cast<long>(data[offset]) << 24) |
-                (static_cast<long>(data[offset + 1]) << 16) |
-                (static_cast<long>(data[offset + 2]) << 8) |
-                static_cast<long>(data[offset + 3]);
+        value = static_cast<long>(
+            (static_cast<uint32_t>(data[offset]) << 24) |
+            (static_cast<uint32_t>(data[offset + 1]) << 16) |
+            (static_cast<uint32_t>(data[offset + 2]) << 8) |
+            static_cast<uint32_t>(data[offset + 3]));
         offset += 4;
         break;
       case MessageSetType::Structure: {
